+++ tests/models.py
@@ -3,10 +3,12 @@ from django.db import models
 from ordered_model.models import OrderedModel, OrderedModelBase
 
 
+# test simple automatic ordering
 class Item(OrderedModel):
     name = models.CharField(max_length=100)
 
 
+# test Answer.order_with_respect_to being a tuple
 class Question(models.Model):
     pass
 
@@ -31,12 +33,14 @@ class Answer(OrderedModel):
         )
 
 
+# test ordering whilst overriding the automatic primary key (ie. not models.Model.id)
 class CustomItem(OrderedModel):
-    id = models.CharField(max_length=100, primary_key=True)
+    pkid = models.CharField(max_length=100, primary_key=True)
     name = models.CharField(max_length=100)
     modified = models.DateTimeField(null=True, blank=True)
 
 
+# test ordering over custom ordering field (ie. not OrderedModel.order)
 class CustomOrderFieldModel(OrderedModelBase):
     sort_order = models.PositiveIntegerField(editable=False, db_index=True)
     name = models.CharField(max_length=100)
@@ -46,13 +50,20 @@ class CustomOrderFieldModel(OrderedModelBase):
         ordering = ("sort_order",)
 
 
+# test ThroughModel ordering with Pizzas/Topping
 class Topping(models.Model):
     name = models.CharField(max_length=100)
+
+    def __str__(self):
+        return self.name
 
 
 class Pizza(models.Model):
     name = models.CharField(max_length=100)
     toppings = models.ManyToManyField(Topping, through="PizzaToppingsThroughModel")
+
+    def __str__(self):
+        return self.name
 
 
 class PizzaToppingsThroughModel(OrderedModel):
@@ -64,6 +75,28 @@ class PizzaToppingsThroughModel(OrderedModel):
         ordering = ("pizza", "order")
 
 
+# Admin only allows each model class to be registered once. However you can register a proxy class,
+# and (for presentation purposes only) rename it to match the existing in Admin
+class PizzaProxy(Pizza):
+    class Meta:
+        proxy = True
+        verbose_name = "Pizza"
+        verbose_name_plural = "Pizzas"
+
+
+# test many-one where the item has custom PK
+class CustomPKGroup(models.Model):
+    name = models.CharField(max_length=100)
+
+
+class CustomPKGroupItem(OrderedModel):
+    group = models.ForeignKey(CustomPKGroup, on_delete=models.CASCADE)
+    name = models.CharField(max_length=100, primary_key=True)
+    order_with_respect_to = "group"
+
+
+# test ordering on a base class (with order_class_path)
+# ie. OpenQuestion and GroupedItem can be ordered wrt each other
 class BaseQuestion(OrderedModel):
     order_class_path = __module__ + ".BaseQuestion"
     question = models.TextField(max_length=100)
@@ -83,6 +116,7 @@ class OpenQuestion(BaseQuestion):
     answer = models.TextField(max_length=100)
 
 
+# test grouping by a foreign model field (group__user)
 class ItemGroup(models.Model):
     user = models.ForeignKey(
         TestUser, on_delete=models.CASCADE, related_name="item_groups"