+++ tests/tests.py
@@ -1,8 +1,11 @@
 import uuid
+from io import StringIO
 
 from django.contrib.auth.models import User
+from django.core.management import call_command
 from django.utils.timezone import now
 from django.test import TestCase
+from django import VERSION
 
 from tests.models import (
     Answer,
@@ -10,6 +13,8 @@ from tests.models import (
     Question,
     CustomItem,
     CustomOrderFieldModel,
+    CustomPKGroupItem,
+    CustomPKGroup,
     Pizza,
     Topping,
     PizzaToppingsThroughModel,
@@ -273,10 +278,10 @@ class OrderWithRespectToTests(TestCase):
 
 class CustomPKTest(TestCase):
     def setUp(self):
-        self.item1 = CustomItem.objects.create(id=str(uuid.uuid4()), name="1")
-        self.item2 = CustomItem.objects.create(id=str(uuid.uuid4()), name="2")
-        self.item3 = CustomItem.objects.create(id=str(uuid.uuid4()), name="3")
-        self.item4 = CustomItem.objects.create(id=str(uuid.uuid4()), name="4")
+        self.item1 = CustomItem.objects.create(pkid=str(uuid.uuid4()), name="1")
+        self.item2 = CustomItem.objects.create(pkid=str(uuid.uuid4()), name="2")
+        self.item3 = CustomItem.objects.create(pkid=str(uuid.uuid4()), name="3")
+        self.item4 = CustomItem.objects.create(pkid=str(uuid.uuid4()), name="4")
 
     def test_saved_order(self):
         self.assertSequenceEqual(
@@ -571,14 +576,33 @@ class OrderedModelAdminTest(TestCase):
         self.assertEqual(Item.objects.get(name="item3").order, 1)
 
     def test_move_up_down_links_ordered_inline(self):
+        # model list
         res = self.client.get("/admin/tests/pizza/")
-        self.assertEqual(res.status_code, 200)
+        self.assertContains(
+            res, text="/admin/tests/pizza/{}/change/".format(self.pizza.id)
+        )
+
+        # model page including inlines
+        res = self.client.get("/admin/tests/pizza/{}/change/".format(self.pizza.id))
+        self.assertContains(
+            res,
+            text='<a href="/admin/tests/pizza/{}/pizzatoppingsthroughmodel/{}/move-up/">'.format(
+                self.pizza.id, self.pizza_to_ham.id
+            ),
+        )
+        self.assertContains(
+            res,
+            text='<a href="/admin/tests/pizza/{}/pizzatoppingsthroughmodel/{}/move-up/">'.format(
+                self.pizza.id, self.pizza_to_pineapple.id
+            ),
+        )
+
+        # click the move-up link
         self.assertEqual(self.pizza_to_ham.order, 0)
         self.assertEqual(self.pizza_to_pineapple.order, 1)
-
         res = self.client.get(
             "/admin/tests/pizza/{}/pizzatoppingsthroughmodel/{}/move-up/".format(
-                self.pizza.id, self.pineapple.id
+                self.pizza.id, self.pizza_to_pineapple.id
             ),
             follow=True,
         )
@@ -587,6 +611,22 @@ class OrderedModelAdminTest(TestCase):
         self.assertEqual(self.pizza_to_ham.order, 1)
         self.assertEqual(self.pizza_to_pineapple.order, 0)
         self.assertEqual(res.status_code, 200)
+
+    def test_move_up_down_proxy_stacked_inline(self):
+        res = self.client.get("/admin/tests/pizzaproxy/")
+        self.assertContains(
+            res, text="/admin/tests/pizzaproxy/{}/change/".format(self.pizza.id)
+        )
+
+        res = self.client.get(
+            "/admin/tests/pizzaproxy/{}/change/".format(self.pizza.id)
+        )
+        self.assertContains(
+            res,
+            text='<a href="/admin/tests/pizzaproxy/{}/pizzatoppingsthroughmodel/{}/move-up/">'.format(
+                self.pizza.id, self.pizza_to_ham.id
+            ),
+        )
 
 
 class OrderWithRespectToTestsManyToMany(TestCase):
@@ -841,8 +881,8 @@ class OrderWithRespectToRelatedModelFieldTests(TestCase):
         )
 
 
-class PolymorpicOrderGenerationTests(TestCase):
-    def test_order_of_Baselist(self):
+class PolymorphicOrderGenerationTests(TestCase):
+    def test_order_of_baselist(self):
         o1 = OpenQuestion.objects.create()
         self.assertEqual(o1.order, 0)
         o1.save()
@@ -864,6 +904,10 @@ class PolymorpicOrderGenerationTests(TestCase):
         self.assertEqual(o2.order, 2)
         m1.refresh_from_db()
         self.assertEqual(m1.order, 3)
+
+    def test_returns_polymorphic(self):
+        o1 = OpenQuestion.objects.create()
+        self.assertIsInstance(o1, OpenQuestion)
 
 
 class BulkCreateTests(TestCase):
@@ -920,4 +964,153 @@ class BulkCreateTests(TestCase):
                 "order", flat=True
             ),
             [0, 1],
+        )
+
+
+class OrderedModelAdminWithCustomPKInlineTest(TestCase):
+    def setUp(self):
+        User.objects.create_superuser("admin", "a@example.com", "admin")
+        self.assertTrue(self.client.login(username="admin", password="admin"))
+        group = CustomPKGroup.objects.create(name="g1")
+        CustomPKGroupItem.objects.create(name="g1 i1", group=group)
+        CustomPKGroupItem.objects.create(name="g1 i2", group=group)
+        group = CustomPKGroup.objects.create(name="g2")
+        CustomPKGroupItem.objects.create(name="g2 i1", group=group)
+
+    def test_move_links(self):
+        res = self.client.get("/admin/tests/custompkgroup/1/change", follow=True)
+        self.assertContains(res, text="CustomPKGroupItem object (g1 i1)")
+        self.assertContains(res, text="CustomPKGroupItem object (g1 i2)")
+
+        # Check for the inline column header
+        # see Django release notes https://docs.djangoproject.com/en/dev/releases/2.2/#django-contrib-admin
+        # What’s new in Django 2.2 > Minor features > django.contrib.admin > Addd a CSS class to the column headers of TabularInline
+        if VERSION >= (2, 2):
+            self.assertContains(
+                res, text='<th class="column-move_up_down_links">Move</th>', html=True
+            )
+        else:
+            self.assertContains(
+                res, text="<th>Move</th>", html=True
+            )  # pragma: no cover
+
+        # Check move up/down links
+        self.assertContains(
+            res,
+            text='<a href="/admin/tests/custompkgroup/1/custompkgroupitem/g1%20i1/move-up/">',
+        )
+        self.assertContains(
+            res,
+            text='<a href="/admin/tests/custompkgroup/1/custompkgroupitem/g1%20i1/move-down/">',
+        )
+
+
+class ReorderModelTestCase(TestCase):
+    fixtures = ["test_items.json"]
+
+    def test_reorder_with_no_respect_to(self):
+        """
+        Test that 'reorder_model' changes the order of OpenQuestions
+        when they overlap.
+        """
+        OpenQuestion.objects.create(order=0)
+        OpenQuestion.objects.create(order=0)
+        out = StringIO()
+        call_command("reorder_model", "tests.OpenQuestion", verbosity=1, stdout=out)
+
+        self.assertSequenceEqual(
+            OpenQuestion.objects.values_list("order", flat=True).order_by("order"),
+            [0, 1],
+        )
+        self.assertIn(
+            "changing order of tests.OpenQuestion (2) from 0 to 1", out.getvalue()
+        )
+
+    def test_reorder_with_respect_to(self):
+        """
+        Test that when 'with_respect_to' is used 'reorder_model' changes to
+        values of the 'order' field to unique values.
+        """
+        user1 = TestUser.objects.create()
+        group1 = ItemGroup.objects.create(user=user1)
+
+        GroupedItem.objects.create(group=group1, order=0)
+        GroupedItem.objects.create(group=group1, order=1)
+        GroupedItem.objects.create(group=group1, order=1)
+        GroupedItem.objects.create(group=group1, order=3)
+        GroupedItem.objects.create(group=group1, order=4)
+
+        user2 = TestUser.objects.create()
+        group2 = ItemGroup.objects.create(user=user2)
+
+        GroupedItem.objects.create(group=group2)
+        GroupedItem.objects.create(group=group2)
+        GroupedItem.objects.create(group=group2)
+
+        out = StringIO()
+        call_command("reorder_model", "tests.GroupedItem", verbosity=1, stdout=out)
+
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group=group1)
+            .values_list("order", flat=True)
+            .order_by("order"),
+            [0, 1, 2, 3, 4],
+        )
+
+        self.assertSequenceEqual(
+            GroupedItem.objects.filter(group=group2)
+            .values_list("order", flat=True)
+            .order_by("order"),
+            [0, 1, 2],
+        )
+
+        self.assertEquals(
+            "changing order of tests.GroupedItem (3) from 1 to 2\n", out.getvalue()
+        )
+
+    def test_reorder_with_custom_order_field(self):
+        """
+        Test that 'reorder_model' changes the order of OpenQuestions
+        when they overlap.
+        """
+        out = StringIO()
+        call_command(
+            "reorder_model", "tests.CustomOrderFieldModel", verbosity=1, stdout=out
+        )
+
+    def test_shows_alternatives(self):
+        out = StringIO()
+        call_command("reorder_model", "test.Missing", verbosity=1, stdout=out)
+        self.assertIn("Model 'test.Missing' is not an ordered model", out.getvalue())
+        self.assertIn("tests.BaseQuestion", out.getvalue())
+
+        out = StringIO()
+        call_command("reorder_model", verbosity=1, stdout=out)
+        self.assertIn("tests.BaseQuestion", out.getvalue())
+
+    def test_delete_bypass(self):
+        OpenQuestion.objects.create(answer="1", order=0)
+        OpenQuestion.objects.create(answer="2", order=1)
+        OpenQuestion.objects.create(answer="3", order=2)
+        OpenQuestion.objects.create(answer="4", order=3)
+
+        # bypass our OrderedModel delete logic to leave a hole in ordering
+        OpenQuestion.objects.filter(answer="3").delete()
+
+        self.assertEqual([0, 1, 3], [i.order for i in OpenQuestion.objects.all()])
+        self.assertEqual(
+            ["1", "2", "4"], [i.answer for i in OpenQuestion.objects.all()]
+        )
+
+        # repair
+        out = StringIO()
+        call_command("reorder_model", "tests.OpenQuestion", stdout=out)
+
+        self.assertEqual([0, 1, 2], [i.order for i in OpenQuestion.objects.all()])
+        self.assertEqual(
+            ["1", "2", "4"], [i.answer for i in OpenQuestion.objects.all()]
+        )
+
+        self.assertEquals(
+            "changing order of tests.OpenQuestion (4) from 3 to 2\n", out.getvalue()
         )