+++ CHANGES.md
@@ -1,18 +1,49 @@
 Change log
 ==========
 
+3.4.3 - 2021-04-20
+------------------
+
+- Fix packaging, setup.py was missing management command package path
+
+
+3.4.2 - 2021-04-20
+------------------
+
+- Fix `OrderedTabularInline` for models with custom primary key field (#233)
+- Add management command `reorder_model` that can re-order most models with a broken ordering (#240)
+- Fix handling of keyword arguments passed to  `bulk_create` by Django 3 (#235)
+- Fix inline admin support for Proxy Models by adding parent model to url name (#242)
+- Migrated to GitHub Actions workflow (#241)
+
+3.4.1 - 2020-05-11
+------------------
+
+- Fix regression in admin OrderedInlineMixin after refactor in 3.4.0
+
+3.4.0 - 2020-05-11
+------------------
+
+- Fix `bulk_create` not returning
+- Fix `OrderedModelQuerySet` returning parent class instances for polymorphic cases
+- Support django 3.0
+- Drop support python 3.4
+
 3.3.0 - 2019-07-10
 ------------------
+
 - `bulk_create` now works with `order_with_respect_to`
 - more internal refactoring moved most methods to `OrderedModelQuerySet`
 
 3.2.0 - 2019-07-10
 ------------------
+
 - Internal refactoring now using `Manager` - probably will break some code
 - provide `bulk_create`
 
 3.1.1 - 2018-11-13
 ------------------
+
 - Fix arrow-top and arrow-bottom not found
 
 3.1.0 - 2018-11-10
@@ -39,8 +70,8 @@ Change log
 ------------------
 
 - Add support for Django 2.1
-- Support order_with_respect_to on related fields
-- Add Tabular and Stacked inline 
+- Support `order_with_respect_to` on related fields
+- Add Tabular and Stacked inline
 
 2.0.0 - 2018-06-07
 ------------------